+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Updated
   * AZ-45: Revamp module
 
 Changed
-  * AZ-45: Use for_each instead of count on firewall_rules / postgre_configuration / vnet_rules
+  * AZ-45: Use for_each instead of count on resources
 
 # v1.0.0 - 2019-10-18